EPL-5:  - Added addTeamCommand

+++ src/main/java/com/github/EPLBot/repository/TelegramUserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TelegramUserRepository extends JpaRepository<TelegramUser, Integer> {
+public interface TelegramUserRepository extends JpaRepository<TelegramUser, Long> {
     List<TelegramUser> findAllByActiveTrue();
 }
